requirements.txt
Django==3.1.7
python-dotenv==0.15.0
django-connection-url==0.1.2
psycopg2==2.9.3
requests==2.25.1
gunicorn==19.9.0
whitenoise==6.2.0
Markdown==3.4.1
djangorestframework==3.12.2